backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
motor==3.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
